+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-﻿### 4.1.39.0 (2022-03-09 00:15Z)
+﻿### 4.1.40.0 (2022-03-09 23:12Z)
+  * AWS Tools for PowerShell now use AWS .NET SDK 3.7.226.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
+  * Amazon Comprehend
+    * Added cmdlet Get-COMPTargetedSentimentDetectionJob leveraging the DescribeTargetedSentimentDetectionJob service API.
+    * Added cmdlet Get-COMPTargetedSentimentDetectionJobList leveraging the ListTargetedSentimentDetectionJobs service API.
+    * Added cmdlet Start-COMPTargetedSentimentDetectionJob leveraging the StartTargetedSentimentDetectionJob service API.
+    * Added cmdlet Stop-COMPTargetedSentimentDetectionJob leveraging the StopTargetedSentimentDetectionJob service API.
+
+### 4.1.39.0 (2022-03-09 00:15Z)
   * AWS Tools for PowerShell now use AWS .NET SDK 3.7.225.0 and leverage its new features and improvements. Please find a description of the changes at https://github.com/aws/aws-sdk-net/blob/master/SDK.CHANGELOG.md.
 
 ### 4.1.38 (2022-03-07 22:06Z)